pancake deployment

+++ test/foundry/src/interfaces/external/univ3/pool/IUniswapV3PoolState.sol
@@ -27,7 +27,7 @@ interface IUniswapV3PoolState {
             uint16 observationIndex,
             uint16 observationCardinality,
             uint16 observationCardinalityNext,
-            uint8 feeProtocol,
+            uint32 feeProtocol,
             bool unlocked
         );
